==> code/notebooks/01-loadData.ipynb <==
import numpy as np
import pandas as pd
import dtale
import pandas_profiling as pp
from IPython.display import Javascript

# load accidents victims dataset

accident_victims_2019_df = pd.read_excel('../../data/bronze/TF_ACCIDENTS_VICTIMS_2019.xlsx')

accident_victims_2020_df = pd.read_excel('../../data/bronze/TF_ACCIDENTS_VICTIMS_2020.xlsx')

# load accidents only dataset

accident_2019_df = pd.read_excel('../../data/bronze/TF_ACCIDENTS_2019.xlsx')

accident_2020_df = pd.read_excel('../../data/bronze/TF_ACCIDENTS_2020.xlsx')

# load air pollution dataset

# check first 2 lines
! head -n 2 ../../data/bronze/Air_pollution.csv

air_pollution_df = pd.read_csv('../../data/bronze/Air_pollution.csv', delimiter=';')

air_pollution_df

# load osm dataset

osm_df = pd.read_excel('../../data/bronze/osm_data.xlsx')

# load bike stations dataset 

bike_stations_df = pd.read_json('../../data/bronze/stations_villo.json')

# load village population statistics

**alternative**

village_populations_wiki_df = pd.read_html('https://nl.wikipedia.org/wiki/Tabel_van_Belgische_gemeenten', thousands='.')

village_populations_wiki_df = village_populations_wiki_df[0][["Gemeente", "Inwoners", "Opp.(km²)"]].reset_index(drop=True)

village_populations_wiki_df.columns = [['Village', '1846', '1990', '1947', '2000', '2021', 'kmOpp']]

village_populations_wiki_df = village_populations_wiki_df[["Village", "2000", "kmOpp"]]

village_populations_wiki_df.columns = ["Village", "Population", "km2_size"]

village_populations_wiki_df

# Convert columns to best possible dtypes

accident_victims_2019_df = accident_victims_2019_df.convert_dtypes()
accident_victims_2020_df = accident_victims_2020_df.convert_dtypes()
accident_2019_df = accident_2019_df.convert_dtypes()
accident_2020_df = accident_2020_df.convert_dtypes()
air_pollution_df = air_pollution_df.convert_dtypes()
osm_df = osm_df.convert_dtypes()
bike_stations_df = bike_stations_df.convert_dtypes()
# village_populations_wiki_df = village_populations_wiki_df.convert_dtypes()

# data saving into silver

# we loaded and cleaned some data and store it from bronze into silver folder

# export without the index, add encoding, delimiter and rounded to two decimals
accident_victims_2019_df.to_csv("../../data/silver/accident_victims_2019_df.csv", encoding='utf-8', sep=';', index=False)
accident_victims_2020_df.to_csv("../../data/silver/accident_victims_2020_df.csv", encoding='utf-8', sep=';', index=False)
accident_2019_df.to_csv("../../data/silver/accident_2019_df.csv", encoding='utf-8', sep=';', index=False)
accident_2020_df.to_csv("../../data/silver/accident_2020_df.csv", encoding='utf-8', sep=';', index=False)
air_pollution_df.to_csv("../../data/silver/air_pollution_df.csv", encoding='utf-8', sep=';', index=False)
osm_df.to_csv("../../data/silver/osm_df.csv", encoding='utf-8', sep=';', index=False)
bike_stations_df.to_csv("../../data/silver/bike_stations_df.csv", encoding='utf-8', sep=';', index=False)
village_populations_wiki_df.to_csv("../../data/silver/village_populations_wiki_df.csv", encoding='utf-8', sep=';', index=False)

# parquet preserve the schema with optimal compression
accident_victims_2019_df.to_parquet("../../data/silver/accident_victims_2019_df.parquet")
accident_victims_2020_df.to_parquet("../../data/silver/accident_victims_2020_df.parquet")
accident_2019_df.to_parquet("../../data/silver/accident_2019_df.parquet")
accident_2020_df.to_parquet("../../data/silver/accident_2020_df.parquet")
air_pollution_df.to_parquet("../../data/silver/air_pollution_df.parquet")
# osm_df.to_parquet("../../data/silver/osm_df.parquet")
bike_stations_df.to_parquet("../../data/silver/bike_stations_df.parquet")
village_populations_wiki_df.to_parquet("../../data/silver/village_populations_wiki_df.parquet")
