==> code/notebooks/02-accidentRatio_bicycles.ipynb <==
import numpy as np
import pandas as pd
import dtale
import pandas_profiling as pp
from IPython.display import Javascript
from plotly.offline import iplot, init_notebook_mode

# load the dataset
accident_victims_2019_df = pd.read_parquet("../../data/silver/accident_victims_2019_df.parquet")

accident_victims_2019_df.info()

# remove any duplicate values
accident_victims_2019_df = accident_victims_2019_df.drop_duplicates().reset_index(drop=True)

# keep only the victim type 'bicycle' users

# remove all other records, keep only the victim type 'bicycle' type 
accident_victims_2019_df = accident_victims_2019_df[(accident_victims_2019_df['TX_VICT_TYPE_DESCR_NL'] == 'Fietser')]

# select only a few important columns (date, vilage, ...)
accident_per_day_2019_df = accident_victims_2019_df[["DT_DAY", "TX_BUILD_UP_AREA_DESCR_NL", "CD_ROAD_USER_TYPE", "TX_ROAD_TYPE_DESCR_NL", "TX_LIGHT_COND_DESCR_NL", "TX_AGE_CLS_DESCR_NL", "TX_MUNTY_DESCR_NL"]]

accident_per_day_2019_df.info()

accident_per_day_2019_df

## count is the right function to use
but, we need the number of incidents that day

# sort date first, then sort on vilagename, then group these two and count() the number of incidents that occured
# then create a new dataframe and sort the results with the most accidents on a given day first 
groupby_accident_per_day_2019_df = accident_per_day_2019_df.groupby(['DT_DAY', 'TX_MUNTY_DESCR_NL'])['TX_ROAD_TYPE_DESCR_NL'].count().to_frame(name='nr_of_accidents').reset_index()
groupby_accident_per_day_2019_df = groupby_accident_per_day_2019_df.sort_values(by=['nr_of_accidents', 'DT_DAY', 'TX_MUNTY_DESCR_NL'], ascending=[False, True, False])
groupby_accident_per_day_2019_df

# this was to check and verify the correctness of the results - ALL GOOD
groupby_accident_per_day_2019_df[groupby_accident_per_day_2019_df["TX_MUNTY_DESCR_NL"] == 'Antwerpen'].sort_values(by=['DT_DAY'], ascending=[True])

# so now, we calculate the accidents/cyclists RATIO

# load population values
village_populations_wiki_df = pd.read_parquet("../../data/silver/village_populations_wiki_df.parquet")

# village_populations_wiki_df.columns = ["Village", "Population", "km2_size"]

village_populations_wiki_df

village_populations_wiki_df['km2_size'] = village_populations_wiki_df['km2_size'].str[:-3:].astype(int)

village_populations_wiki_df.info()

# now we calculate the ratio

groupby_accident_per_day_2019_df = groupby_accident_per_day_2019_df.replace({'TX_MUNTY_DESCR_NL':{
    'Sint-Niklaas (Sint-Niklaas)': 'Sint-Niklaas',
    'Halle (Halle-Vilvoorde)': 'Halle',
    'Beveren (Sint-Niklaas)': 'Beveren',
    'Tielt (Tielt)': 'Tielt',
    'Vorst (Brussel-Hoofdstad)': 'Vorst',
    'Kapellen (Antwerpen)': 'Kapellen',
    'Hamme (Dendermonde)': 'Hamme',
    'Nieuwerkerken (Hasselt)': 'Nieuwerkerken',
    'Machelen (Halle-Vilvoorde)': 'Machelen',
    'Hove (Antwerpen)': 'Hove',
    'Celles (Doornik)': 'Celles',
    'Perwijs (Nijvel)': 'Perwijs',
    'Saint-Nicolas (Luik)': 'Saint-Nicolas',
    'Aalst (Aalst)': 'Aalst',
    'Moerbeke (Gent)': 'Moerbeke'}})

village_populations_wiki_df = village_populations_wiki_df.replace({'Village':{
    'Pelt*': 'Pelt',
    'Deinze*': 'Deinze',
    'Kruisem*': 'Kruisem',
    'Aalter*': 'Aalter',
    'Belœil': 'Beloeil',
    'Lievegem*': 'Lievegem',
    "Mont-de-l'Enclus": 'Mont-de-l’Enclus',
    'Puurs-Sint-Amands*': 'Puurs-Sint-Amands',
    'Court-Saint-Étienne': 'Court-Saint-Etienne',
    'Éghezée': 'Eghezée',
    'Oudsbergen*': 'Oudsbergen',
    'Estaimpuis*': 'Estaimpuis',
    "Fontaine-l'Evêque": 'Fontaine-l’Evêque',
    'Gembloers': 'Gembloux',
    '’s Gravenbrakel*': '’s Gravenbrakel',
    'Erezée*': 'Erezée',
    'Ecaussines': 'Ecaussinnes',
    'Blegny*': 'Blegny',
    'Plombières*': 'Plombières'}})

# merge join on village
merged_accidents_df = groupby_accident_per_day_2019_df.merge(village_populations_wiki_df, left_on='TX_MUNTY_DESCR_NL', right_on='Village', how='left')

# remove Nan records
merged_accidents_df = merged_accidents_df[~(merged_accidents_df['Population'].isna())]

# calculate the ratio itself

### (nr of incidents / population ) * city size in km

merged_accidents_df.loc[:, 'accident_village_ratio'] = merged_accidents_df['nr_of_accidents'] / merged_accidents_df['Population']

merged_accidents_df.loc[:, 'accident_village_size_ratio'] = merged_accidents_df['accident_village_ratio'] * merged_accidents_df['km2_size']

merged_accidents_df

accident_village_ratio_median_cutoff = merged_accidents_df['accident_village_ratio'].median()
accident_village_ratio_median_cutoff

accident_village_size_ratio_median_cutoff = merged_accidents_df['accident_village_size_ratio'].median()
accident_village_size_ratio_median_cutoff

# chart_data = pd.concat([
#     merged_accidents_df['Village'],
#     merged_accidents_df['accident_village_size_ratio'],
# ], axis=1)
# chart_data = chart_data.sort_values(['Village'])
# chart_data = chart_data.rename(columns={'Village': 'x'})
# chart_data = chart_data.dropna()

# import plotly.graph_objs as go

# charts = []
# charts.append(go.Bar(
#     x=chart_data['x'],
#     y=chart_data['accident_village_size_ratio']
# ))
# figure = go.Figure(data=charts, layout=go.Layout({
#     'barmode': 'stack',
#     'legend': {'orientation': 'h'},
#     'title': {'text': 'accident_village_size_ratio by Village'},
#     'xaxis': {'title': {'text': 'Village'}},
#     'yaxis': {'title': {'text': 'accident_village_size_ratio'}, 'type': 'linear'}
# }))

# from plotly.offline import iplot, init_notebook_mode

# init_notebook_mode(connected=True)
# for chart in charts:
#     chart.pop('id', None) # for some reason iplot does not like 'id'
# iplot(figure)

# # generate a data report to get a feel of the data (EDA)
# d = dtale.show(merged_accidents_df, host="20.223.36.46", port="40000", ignore_duplicate=True, drop_index=True, reaper_on=False)
# # show all running instances
# d.main_url()

result = merged_accidents_df.groupby(['Village'])['accident_village_size_ratio'].sum().to_frame(name='ratio').reset_index()
result = result.sort_values(by=['ratio', 'Village'], ascending=[False, True])
result.head(10)

# generate a data report to get a feel of the data (EDA)
d = dtale.show(result, host="20.223.36.46", port="40000", ignore_duplicate=True, drop_index=True, reaper_on=False)
# show all running instances
d.main_url()

s = result['ratio']
q1 = s.quantile(0.25)
q3 = s.quantile(0.75)
iqr = q3 - q1
iqr_lower = q1 - 1.5 * iqr
iqr_upper = q3 + 1.5 * iqr
outliers = dict(s[(s < iqr_lower) | (s > iqr_upper)])

result.iloc[list(outliers.keys())]

chart_data = pd.concat([
    result.iloc[list(outliers.keys())]['Village'],
    result.iloc[list(outliers.keys())]['ratio'],
], axis=1)
chart_data = chart_data.sort_values(['Village'])
chart_data = chart_data.rename(columns={'Village': 'x'})
chart_data = chart_data.dropna()

import plotly.graph_objs as go

charts = []
charts.append(go.Bar(
    x=chart_data['x'],
    y=chart_data['ratio']
))
figure = go.Figure(data=charts, layout=go.Layout({
    'barmode': 'stack',
    'legend': {'orientation': 'h'},
    'title': {'text': 'Outliers (highest & lowest)'},
    'xaxis': {'title': {'text': 'Cities'}},
    'yaxis': {'title': {'text': 'ratio of cylists'}, 'type': 'linear'}
}))

init_notebook_mode(connected=True)
for chart in charts:
    chart.pop('id', None) # for some reason iplot does not like 'id'
iplot(figure)

figure.write_image("../../data/plot/accident_ratio_bicycles.png")

reportFile = "../../data/report/accident_ratio_bicycles_report.html"

# quick report on 100% records (no correlation matrix stuff)
pp.ProfileReport(df=result.sample(frac=1),
                 minimal=True,
                 progress_bar=False,
                 correlations={"cramers": {"calculate": False}},
                 title="Report Title",
                 html={"style": {"full_width": True}}).to_file(reportFile)

# open the report (*.html)
display(Javascript('window.open("{url}");'.format(url=reportFile)))
