==> code/notebooks/03-airPollution-cities_Kevin.ipynb <==
import pandas as pd
import numpy as np
import random
import folium
from folium.plugins import HeatMap
import matplotlib.pyplot as plt

import numpy as np
import pandas as pd
import dtale
import pandas_profiling as pp
from IPython.display import Javascript

# import module
from geopy.geocoders import Nominatim

airpollute=pd.read_csv('../../data/bronze/Air_pollution.csv', delimiter=';')

airpollute

# split the coordinates
airpollute['lat'], airpollute['lon'] = airpollute["Coordinates"].str.split(',').str

# # generate a data report to get a feel of the data (EDA)
# d = dtale.show(airpollute, host="20.223.36.46", port="40000", ignore_duplicate=True, drop_index=True, reaper_on=False)
# # show all running instances
# d.main_url()

# initialize Nominatim API
geolocator = Nominatim(user_agent="geoapiExercises")

# # Latitude & Longitude input
# Latitude = "51.023378"
# Longitude = "4.475089"

# location = geolocator.reverse(Latitude+","+Longitude)

# location

# Display
# location.address.split(',')

# for coordinates in airpollute['Coordinates'].str.split(',')[0]:

location_str = []
for i in airpollute['Coordinates'].str.split(','):
    location_str.append(geolocator.reverse(i[0]+","+i[1]))
    print(i)

# location_str[0][0].split(',')[3].split(' - ')[-1].strip()

city = []
for i in range(len(airpollute)):
    city.append(location_str[i][0].split(',')[3].split(' - ')[-1].strip())
city

airpollute['Village'] = city

airpollute

airpollute.columns = ['Region', 'Location', 'Coordinates', 'CountryCode', 'Pollutant', 'SourceName', 'Unit', 'Value', 'LastUpdated', 'Country', 'lat', 'lon', 'City']

# airpollute = airpollute[['Pollutant', 'Unit', 'Value', 'lon', 'lat', 'City']]

airpollute.info()

airpollute = airpollute.convert_dtypes()

airpollute.info()

# store/save the current results
airpollute.to_csv("../../data/silver/cleaned_air_pollution_df.csv", encoding='utf-8', sep=';', index=False)

# load the dataset

airpollute = pd.read_csv("../../data/silver/cleaned_air_pollution_df.csv", encoding='utf-8', sep=';')

# generate a data report to get a feel of the data (EDA)
d = dtale.show(airpollute, host="20.223.36.46", port="40000", ignore_duplicate=True, drop_index=True, reaper_on=False)
# show all running instances
d.main_url()

airpo=airpollute[['City', 'Location','Pollutant' , 'Value','Coordinates', 'Country', 'Last Updated']]

long_lat=airpo.Coordinates.str.split(',', expand = True)
airpo['lat']=long_lat[0]
airpo['long']=long_lat[1]
airpo=airpo[['City', 'Location','Pollutant' , 'Value','long', 'lat','Country', 'Last Updated']]
#经度在前

airpo.head()
# date handling
airpo['Date'] = pd.to_datetime(airpo['Last Updated'], format='%Y-%m-%d')
airpo['Year'] = list(map(lambda x: int(x.year), airpo.Date))

airpo.head()

a=airpo[['City', 'long', 'lat']]
long_lat=a.groupby('City', as_index=False).first()
long_lat

# count the pollutant
pollute=list(airpo.Pollutant.unique())
pollute_counts=airpo.Pollutant.value_counts()
pollute_counts

# pollution ratio
city_pollute=airpo.groupby(['City'],as_index=False).Value.sum()
city_pollute['long']=long_lat.long
city_pollute['lat']=long_lat.lat
city_pollute['total']=sum(city_pollute.Value)
city_pollute['pollute_ratio']=city_pollute.Value/city_pollute['total']
city_pollute['Location']=airpo['Location']
city_pollute['year']=airpo['Year']
city_pollute.head()

# pollution ratio 2020
airpo2020=airpo[airpo.Year==2020]
city_p2020=airpo2020.groupby(['City'],as_index=False).Value.sum()
city_p2020['long']=long_lat.long
city_p2020['lat']=long_lat.lat
city_p2020['total']=sum(city_p2020.Value)
city_p2020['pollute_ratio']=city_p2020.Value/city_p2020['total']
city_p2020['Location']=airpo['Location']

city_p2020
# using max as a reference group

#for each type of pollutant, compare the amount of pollution within countries
# for 2020 and after
airpo1=airpo[airpo.Year==2020]
airbypo=airpo1.groupby(['Pollutant', 'City'],as_index=False).first()
Pollutant=list(airpo.Pollutant.unique())
for p in Pollutant:
    plt.figure()
    plt.bar(airbypo.City, airbypo.Value)
    plt.xticks(rotation=270)
    plt.title(p)

plt.bar(city_pollute.City, city_pollute.Value)
plt.xticks(rotation=270)
plt.show()
# average

#bar plot for pollution 2020
city_p2020=city_pollute[city_pollute.year==2020]
plt.bar(city_p2020.City, city_p2020.Value)
plt.xticks(rotation=270)
plt.show()

firms=city_p2020.City
city_p2020
#market_share=city_pollute.pollute_ratio
#plt.pie(market_share,  labels=firms, autopct='%1.1f%%')
#plt.show()

firms=city_p2020.City
market_share=city_p2020.pollute_ratio
plt.pie(market_share,  labels=firms, autopct='%1.1f%%')
plt.show()

#pie graph for all year
firms=city_pollute.City
market_share=city_pollute.pollute_ratio
plt.pie(market_share,  labels=firms, autopct='%1.1f%%')
plt.show()

#heat map 
def generateBaseMap(loc, zoom=9, tiles='OpenStreetMap', crs='ESPG2263'):
    return folium.Map(location=loc,
                   control_scale=True, 
                   zoom_start=zoom,
                   tiles=tiles)

MAP = [city_pollute.lat[0],city_pollute.long[0]] 
  
base_map = generateBaseMap(MAP)

city_pollute['Values']=city_pollute['Value']/100
map_values1 = city_pollute[['lat','long','pollute_ratio']]

data = map_values1.values.tolist()
           
hm = HeatMap(data,gradient={0.009: 'blue', 0.02: 'lime', 0.05: 'yellow', 0.08: 'orange', 0.1: 'red'}, 
                min_opacity=0.05, 
                max_opacity=0.9, 
                radius=25,
                use_local_extrema=False)#.add_to(base_map)

#base_map
base_map.add_child(hm)

# interactive map
def plot_map(df_select, i_df, i_scale):
    zoom_factor = 8 # inital map size
    radius_scaling = i_scale # scaling of bubbles

    my_map = folium.Map(location=[50.5,4], zoom_start=zoom_factor)

    for i in range(0,df_select.shape[0]):
        folium.CircleMarker(
          location=[df_select.iloc[i]['lat'], df_select.iloc[i]['long']],
          radius=np.sqrt(df_select.iloc[i]['Value'])*radius_scaling,
          popup='Location: ' + df_select.iloc[i]['Location'] + '\n Value: ' + str(df_select.iloc[i]['Value']),
          color='red',
          fill=True,
          fill_color='red',
          fill_opacity=0.25,
          weight=1 # stroke width in pixels
       ).add_to(my_map)

    return(my_map)

#for each city, proportion of all kinds of pollution in 2020
my_map = plot_map(city_p2020,i_df=city_p2020, i_scale=1)
my_map
